% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigd.R
\name{ugd_clear}
\alias{ugd_clear}
\title{Clear all unigd plot pages.}
\usage{
ugd_clear(which = dev.cur())
}
\arguments{
\item{which}{Which device (ID).}
}
\value{
Whether there were any pages to remove.
}
\description{
This function will only work after starting a device with \code{\link[=ugd]{ugd()}}.
}
\examples{
ugd()
plot(1, 1)
hist(rnorm(100))
ugd_clear() # Clear all previous plots
hist(rnorm(100))

dev.off()
}
