% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigd.R
\name{ugd_info}
\alias{ugd_info}
\title{unigd device information.}
\usage{
ugd_info(which = dev.cur())
}
\arguments{
\item{which}{Which device (ID).}
}
\value{
List of status variables with the following named items:
\verb{$id}: Server unique ID,
\verb{$version}: unigd and library versions.
}
\description{
Access general information of a unigd graphics device.
This function will only work after starting a device with \code{\link[=ugd]{ugd()}}.
}
\examples{
ugd() # Initialize graphics device
ugd_info() # Get device information
dev.off() # Close device
}
