% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigd.R
\name{ugd_inline}
\alias{ugd_inline}
\title{Inline SVG rendering.}
\usage{
ugd_inline(
  code,
  page = 0,
  page_width = -1,
  page_height = -1,
  zoom = 1,
  renderer = "svg",
  file = NA,
  ...
)
}
\arguments{
\item{code}{Plotting code. See examples for more information.}

\item{page}{Plot page to render. If this is set to \code{0}, the last page will
be selected. Can be set to a numeric plot index or plot ID
(see \code{\link[=ugd_id]{ugd_id()}}).}

\item{page_width}{Width of the plot. If this is set to \code{-1}, the last width
will be selected.}

\item{page_height}{Height of the plot. If this is set to \code{-1}, the last
height will be selected.}

\item{zoom}{Zoom level. (For example: \code{2} corresponds to 200\%, \code{0.5} would
be 50\%.)}

\item{renderer}{Renderer.}

\item{file}{Filepath to save SVG. (No file will be created if this is \code{NA})}

\item{...}{Additional parameters passed to \code{ugd(webserver=FALSE, ...)}}
}
\value{
Rendered SVG string.
}
\description{
Convenience function for quick inline SVG rendering.
This is similar to \code{\link[=ugd_plot]{ugd_plot()}} but the plotting code is specified inline
and an offline unigd graphics device is managed (created and closed)
automatically. Starting a device with \code{\link[=ugd]{ugd()}} is therefore not necessary.
}
\examples{
ugd_inline({
  hist(rnorm(100))
})

s <- ugd_inline({
  plot.new()
  lines(c(0.5, 1, 0.5), c(0.5, 1, 1))
})
cat(s)
}
