% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigd.R
\name{ugd_remove}
\alias{ugd_remove}
\title{Remove a unigd plot page.}
\usage{
ugd_remove(page = 0, which = dev.cur())
}
\arguments{
\item{page}{Plot page to remove. If this is set to \code{0}, the last page will
be selected. Can be set to a numeric plot index or plot ID
(see \code{\link[=ugd_id]{ugd_id()}}).}

\item{which}{Which device (ID).}
}
\value{
Whether the page existed (and thereby was successfully removed).
}
\description{
This function will only work after starting a device with \code{\link[=ugd]{ugd()}}.
}
\examples{
ugd()
plot(1, 1) # page 1
hist(rnorm(100)) # page 2
ugd_remove(page = 1) # remove page 1

dev.off()
}
