% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigd.R
\name{ugd_state}
\alias{ugd_state}
\title{unigd device status.}
\usage{
ugd_state(which = dev.cur())
}
\arguments{
\item{which}{Which device (ID).}
}
\value{
List of status variables with the following named items:
\verb{$hsize}: Plot history size (how many plots are accessible),
\verb{$upid}: Update ID (changes when the device has received new information),
\verb{$active}: Is the device the currently activated device.
}
\description{
Access status information of a unigd graphics device.
This function will only work after starting a device with \code{\link[=ugd]{ugd()}}.
}
\examples{
\dontrun{

ugd()
ugd_state()
plot(1, 1)
ugd_state()

dev.off()
}
}
