% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testgraphic.R
\name{ugd_test_pattern}
\alias{ugd_test_pattern}
\title{Plot a test pattern that can be used to evaluate and compare graphics
devices.}
\usage{
ugd_test_pattern()
}
\value{
Nothing, but a plot is generated.
}
\description{
Plot a test pattern that can be used to evaluate and compare graphics
devices.
}
\examples{
\dontrun{

ugd_test_pattern()
}
}
